replace datetimepicker with full js
