+++ src/main/webapp/templates/celMacros/includeDefaultHeader.vm
@@ -59,8 +59,8 @@ $services.javascript.addExtJSfileOnce(':celJS/bootstrap/bootstrap.min.js', 'file
 $services.javascript.addExtJSfileOnce(':celJS/bootstrap/bootstrap-multiselect.js', 'file')
 
 $services.css.includeCSSPage(':celJS/jquery-datetimepicker/2.5/jquery.datetimepicker.min.css')
-##$services.javascript.addExtJSfileOnce(':celJS/jquery-datetimepicker/2.5/jquery.datetimepicker.min.js', 'file')
-$services.javascript.addExtJSfileOnce(':celJS/jquery-datetimepicker/2.5/jquery.datetimepicker.js', 'file')
+##$services.javascript.addExtJSfileOnce(':celJS/jquery-datetimepicker/2.5/jquery.datetimepicker.full.min.js', 'file')
+$services.javascript.addExtJSfileOnce(':celJS/jquery-datetimepicker/2.5/jquery.datetimepicker.full.js', 'file')
 $services.javascript.addExtJSfileOnce(':celJS/dateTimePicker/generateDateTimePicker.js', 'file')
 
 #if($services.celementsphoto.useImageAnimations())
